+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ### Unreleased
+- Be compatiable with refactored @twreporter/registration package
+- Include client.js and Server.js
+- Be compatible with @twreporter/redux which drop bookmarks related stuffs
+- Import bookmarks related statement (mainly redux actions/reducers) from @twreporter/registration
+- Move Functional Bookmark Icons (add/Remove) back to @twreporter/registration.
 
 ### 2.5.2
 - Bump @twreporter/react-components version to 2.1.11